# projekt/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sugárzási API",
    "version": "1.0.0",
    "description": "API három feladat megoldásához és bejelentkezéshez kötött adatbevitelhez"
  },
  "paths": {
    "/sugarzasihely": {
      "get": {
        "summary": "Adott település sugárzási címei (egyediek)",
        "parameters": [
          {
            "name": "telepules",
            "in": "query",
            "required": true,
            "description": "A település neve (pl. Budapest)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Egyedi címek listája",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          },
          "400": {
            "description": "Hiányzó paraméter"
          }
        }
      }
    },
    "/teljesitmeny": {
      "get": {
        "summary": "Adott helyről sugárzott csatornák és teljesítményük (csökkenő sorrend)",
        "parameters": [
          {
            "name": "adohely",
            "in": "query",
            "required": true,
            "description": "A sugárzás helye (pl. Miskolc)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Csatornák és teljesítmény",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "csatorna": { "type": "string" },
                      "teljesitmeny": { "type": "number" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Hiányzó paraméter"
          }
        }
      }
    },
    "/varosinallomasnev": {
      "get": {
        "summary": "Olyan csatornák nevei, amelyek tartalmazzák az adóhely nevét",
        "responses": {
          "200": {
            "description": "Csatornák listája",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },

    "/login": {
      "post": {
        "summary": "Bejelentkezés felhasználónévvel és jelszóval",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sikeres bejelentkezés",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Hibás hitelesítés" }
        }
      }
    },

    "/addtelepules": {
      "post": {
        "summary": "Település beszúrása (JWT szükséges)",
        "parameters": [
          {
            "name": "telepules",
            "in": "query",
            "required": true,
            "description": "Formátum: név!megye",
            "schema": { "type": "string" }
          }
        ],
        "security": [
           { "bearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Sikeres beszúrás" },
          "400": { "description": "Hibás paraméter" },
          "401": { "description": "Hiányzó token" },
          "403": { "description": "Érvénytelen token" }
        }
      }
    },

    "/addkiosztas": {
      "post": {
        "summary": "Kiosztás beszúrása (Csak Bejelentkezve)",
        "parameters": [
          {
            "name": "telepules",
            "in": "query",
            "required": true,
            "description": "Formátum: frekvencia!csatorna!teljesitmeny!adohely!cim",
            "schema": { "type": "string" }
          }
        ],
        "security": [
           { "bearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Sikeres beszúrás" },
          "400": { "description": "Hibás paraméter" },
          "401": { "description": "Hiányzó token" },
          "403": { "description": "Érvénytelen token" }
        }
      }
    }
  },



  
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}